+++ lib/src/progress_bar.dart
@@ -1,6 +1,5 @@
 import 'package:chewie/src/chewie_progress_colors.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:video_player/video_player.dart';
 
 class VideoProgressBar extends StatefulWidget {
@@ -28,6 +27,7 @@ class VideoProgressBar extends StatefulWidget {
   final bool drawShadow;
 
   @override
+  // ignore: library_private_types_in_public_api
   _VideoProgressBarState createState() {
     return _VideoProgressBarState();
   }
@@ -186,7 +186,7 @@ class _ProgressBarPainter extends CustomPainter {
     );
 
     if (drawShadow) {
-      final shadowPath = Path()
+      final Path shadowPath = Path()
         ..addOval(
           Rect.fromCircle(
             center: Offset(playedPart, baseOffset + barHeight / 2),